% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_Genotype_Warnings.R
\name{BC_Genotype_Warnings}
\alias{BC_Genotype_Warnings}
\title{Backcross genotype warnings; must provide warning allele}
\usage{
BC_Genotype_Warnings(seq_data, genotype_table, warning_allele = "Ref",
  output = "warnings")
}
\value{
A dataframe of warnings
}
\description{
Backcross genotype warnings; must provide warning allele
}
\examples{
\dontrun{
genotype_warnings <- BC_Genotype_Warnings(seq_data, genotype_table, warning_allele="Ref", output="warnings")
}
}
