% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_coding.R
\name{binary_coding}
\alias{binary_coding}
\title{Code genotypes as binary}
\usage{
binary_coding(genotype_warnings2NA, genotype_table)
}
\value{
A dataframe of binary coded genotypes as a slot in the input
}
\description{
Code genotypes as binary
}
\examples{
\dontrun{
genotypes_object <- binary_coding(genotype_warnings2NA, genotype_table)
}
}
\keyword{as}
\keyword{binary}
\keyword{code}
\keyword{genotypes}
