% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO.R
\name{CO}
\alias{CO}
\title{Where crossovers occur per individual with 2 ways to deal with
missing data}
\usage{
CO(indata, naive = FALSE)
}
\arguments{
\item{indata}{this is genotype data frame}

\item{naive}{this takes 2 values: 1) FALSE (default) returns
list with COs distributed by marker distance, and 2) TRUE returns a
list with COs without regard to marker distance (i.e., at the final
non-missing data point in a string of missing gentypes)}
}
\value{
list of COs counted per individual
}
\description{
Where crossovers occur per individual with 2 ways to deal with
missing data
}
\examples{
\dontrun{
CO_df <- CO(indata, naive=FALSE)
}##subset function

}
\keyword{count}
\keyword{crossovers}
\keyword{genotypeR}
