% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2qtl_table.R
\name{convert2qtl_table}
\alias{convert2qtl_table}
\title{write out table for import into rqtl}
\usage{
convert2qtl_table(genotypeR_object,
  temp_cross_for_qtl = "temp_cross_for_qtl.csv")
}
\arguments{
\item{genotypeR_object}{this is a genotypeR object that has had
genotypes coded as binary with binary_coding}

\item{temp_cross_for_qtl}{name of the output file that will be output
into the working directory}
}
\value{
table to disk for input into rqtl
}
\description{
write out table for import into rqtl
}
\examples{
\dontrun{
convert2qtl_table(counted_COs, temp_cross_for_qtl="~/Desktop/temp.csv")
}##subset function

}
\keyword{genotypeR,}
\keyword{rqtl}
