% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{description}
\alias{description}
\alias{description,Gene-method}
\title{Accessor function for description}
\usage{
description(gene)

\S4method{description}{Gene}(gene)
}
\arguments{
\item{gene}{An instance of the Gene class}
}
\value{
The description of the gene.

The Gene object with updated description
}
\description{
Accessor function for description
}
\examples{
library(GenomicRanges)
library(Biostrings)
pcg_example <- new("ProteinCodingGene",
  geneID = "ENSG00000139618",
  symbol = "BRCA2",
  name = "Breast cancer 2",
  description = "Gene implicated in breast cancer.",
  structure = GRanges(
    seqnames = Rle("17"),
    ranges = IRanges(start = 43044295, end = 43125482),
    strand = Rle("+")
  ),
  proteinID = "P51587",
  proteinSequence = AAString("MARKSLEMSIR")
)
description(pcg_example)
}
