% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{Gene-class}
\alias{Gene-class}
\title{Gene}
\value{
No objects are returned because gene class is virtual
}
\description{
Gene Class
}
\details{
This virtual class represents the fundamental attributes for a gene.
}
\section{Slots}{

\describe{
\item{\code{geneID}}{Unique gene identifier}

\item{\code{symbol}}{gene symbol (e.g., HUGO symbol).}

\item{\code{name}}{full gene name}

\item{\code{description}}{Brief gene description}

\item{\code{structure}}{Genomic structure as GRanges}
}}

