% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_GeneConstructors.R
\name{Gene}
\alias{Gene}
\alias{createGene}
\title{Create a Gene object}
\usage{
Gene(geneID, symbol, name, description, structure)
}
\arguments{
\item{geneID}{Unique identifier for the gene.}

\item{symbol}{Standard symbol associated with the gene.}

\item{name}{Full name of the gene.}

\item{description}{Detailed description of the gene.}

\item{structure}{A GRanges object of the genomic structure of the gene.}
}
\value{
A new Gene object.
}
\description{
This function initializes a new Gene object with specified attributes.
}
