% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoTypeR-package.R
\docType{package}
\name{GenoTypeR-package}
\alias{GenoTypeR-package}
\alias{GenoTypeR}
\alias{testduplicate-package}
\title{GenoTypeR: A Package for Gene Data Analysis}
\description{
GenoTypeR provides a set of S4 classes for analyzing gene data,
for various types od genes (protein-coding, microRNA, etc.)
and functions to manipulate and analyze these data structures.
}
\details{
#' \tabular{ll}{
Package: \tab GenoTypeR\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2024-09-04\cr
License: \tab GPL (>=3)\cr
}
}
\section{Overview}{

This package includes:
\itemize{
\item S4 class definitions for different types o genes.
\item Methods to calculate gene product lengths.
\item Methods for accessing and modifying gene properties.
}

@section Important functions:
\describe{
\item{\code{\link{lengthProduct}}}{Calculate the length of a gene product}
}
}

\author{
Maria Claudia Savina\cr
Politecnico di Milano\cr
Maria Claudia Savina\cr
E-Mail: \href{mailto:mariaclaudia.savina@mail.polimi.it}{mariaclaudia.savina@mail.polimi.it}
}
\keyword{internal}
