% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypes_data.R
\docType{data}
\name{genotypes_data}
\alias{genotypes_data}
\title{Genotyping data from the sequenom platform
from markers produced with genotypeR}
\format{An object read in with \code{read_in_sequenom_data}; see \code{\link[genotypeR]{read_in_sequenom_data}}.}
\usage{
data(genotypes_data)
}
\description{
Data from a recombination plasticity experiment in 
Drosophila pseudoobscura. This data is provided to demonstrate the
the use of the genotypeR package
}
\examples{
data(genotypes_data)
head(genotypes_data)
colnames(genotypes_data)

}
\keyword{datasets}
