% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illumina_Genotype_Table.R
\name{illumina_Genoype_Table}
\alias{illumina_Genoype_Table}
\title{Make genotypeR Alt_Ref_Table}
\usage{
illumina_Genoype_Table(tab_delimited_file, flanking_region_length, chromosome)
}
\arguments{
\item{tab_delimited_file}{is a tab delimited AB illumina GoldenGate file}

\item{flanking_region_length}{is the length in bp of the flanking region of the SNP}

\item{chromosome}{is a vector of chromosome names}
}
\value{
data frame useful used in genotypeR
}
\description{
Make genotypeR Alt_Ref_Table
}
\examples{
\dontrun{
test_data <- read_in_illumina_GoldenGate(tab_delimited_file="Noor Plates 1-14__Feb-12-10_FinalReport.txt", flanking_region_length=50, chromosome=rep("chr2", length.out=length(552960)))
illumina_table <- illumina_Genoype_Table(tab_delimited_file="Noor Plates 1-14__Feb-12-10_FinalReport.txt", flanking_region_length=50, chromosome=rep("chr2", length.out=length(552960)))
}
}
\keyword{GoldenGate}
\keyword{illumina}
