% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_GeneFunctions.R
\name{lengthProduct,LongNonCodingRNAGene-method}
\alias{lengthProduct,LongNonCodingRNAGene-method}
\title{Length of the RNA product for LongNonCodingRNAGene.}
\usage{
\S4method{lengthProduct}{LongNonCodingRNAGene}(gene)
}
\arguments{
\item{gene}{A \code{LongNonCodingRNAGene} object}
}
\value{
Integer representing the length of the RNA sequence.
}
\description{
Length of the RNA product for LongNonCodingRNAGene.
}
\examples{
library(GenomicRanges)
library(Biostrings)
lnc_example <- new("LongNonCodingRNAGene",
  geneID = "ENSG00000228630",
  symbol = "HOTAIR",
  name = "HOX transcript antisense RNA",
  description = "LncRNA implicated in gene silencing.",
  structure = GRanges(
    seqnames = Rle("12"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("-")
  ),
  lncRNAID = "NR_00316",
  RNASequence = RNAString("UGAGGUAGUAGGUU")
)
lnc_length <- lengthProduct(lnc_example)
print(lnc_length)
}
