% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_GeneFunctions.R
\name{lengthProduct,MicroRNAGene-method}
\alias{lengthProduct,MicroRNAGene-method}
\title{Length of the RNA sequence for MicroRNAGene.}
\usage{
\S4method{lengthProduct}{MicroRNAGene}(gene)
}
\arguments{
\item{gene}{A MicroRNAGene object}
}
\value{
Integer representing the length of the RNA sequence.
}
\description{
Length of the RNA sequence for MicroRNAGene.
}
\examples{
library(GenomicRanges)
library(Biostrings)
miRNA_example <- new("MicroRNAGene",
  geneID = "ENSG00000207734",
  symbol = "MIR21",
  name = "MicroRNA 21",
  description = "miRNA involved in cancer regulation and expression.",
  structure = GRanges(
    seqnames = Rle("17"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("+")
  ),
  microRNAID = "MI000077",
  seedSequence = RNAString("UAGCUUAUCAGAC")
)
miRNA_length <- lengthProduct(miRNA_example)
print(miRNA_length)
}
