% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_GeneFunctions.R
\name{lengthProduct,PiwiInteractingRNAGene-method}
\alias{lengthProduct,PiwiInteractingRNAGene-method}
\title{Length of the piRNA sequence for Piwi-interactingRNAGene.}
\usage{
\S4method{lengthProduct}{PiwiInteractingRNAGene}(gene)
}
\arguments{
\item{gene}{A PiwiInteractingRNAGene object}
}
\value{
Integer representing the length of the piRNA sequence.
}
\description{
Length of the piRNA sequence for Piwi-interactingRNAGene.
}
\examples{
library(GenomicRanges)
library(Biostrings)
piRNA_example <- new("PiwiInteractingRNAGene",
  geneID = "ENSG00000241599",
  symbol = "PIWI001",
  name = "Piwi-interacting RNA 1",
  description = "Piwi-interacting RNA involved in gene regulation",
  structure = GRanges(
    seqnames = Rle("x"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("+")
  ),
  piRNAID = "PIR0001",
  piSequence = RNAString("UGAGGUAGUAGGUU")
)
piRNA_length <- lengthProduct(piRNA_example)
print(piRNA_length)
}
