% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_GeneFunctions.R
\name{lengthProduct,ProteinCodingGene-method}
\alias{lengthProduct,ProteinCodingGene-method}
\title{Product length for Protein Coding Gene}
\usage{
\S4method{lengthProduct}{ProteinCodingGene}(gene)
}
\arguments{
\item{gene}{A ProteinCodingGene object}
}
\value{
Integer representing the length of the protein sequence.
}
\description{
Product length for Protein Coding Gene
}
\examples{
library(GenomicRanges)
library(Biostrings)
pcg_example <- new("ProteinCodingGene",
  geneID = "ENSG00000139618",
  symbol = "BRCA2",
  name = "Breast cancer 2",
  description = "Gene implicated in breast cancer.",
  structure = GRanges(
    seqnames = Rle("17"),
    ranges = IRanges(start = 43044295, end = 43125482),
    strand = Rle("+")
  ),
  proteinID = "P51587",
  proteinSequence = AAString("MVLSPADKTNVK")
)
pcg_length <- lengthProduct(pcg_example)
print(pcg_length)
}
