% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_GeneFunctions.R
\name{lengthProduct,RibosomialRNAGene-method}
\alias{lengthProduct,RibosomialRNAGene-method}
\title{Length of the rRNA sequence for RibosomialRNAGene.}
\usage{
\S4method{lengthProduct}{RibosomialRNAGene}(gene)
}
\arguments{
\item{gene}{A RibosomialRNAGene`object}
}
\value{
Integer representing the length of the rRNA sequence.
}
\description{
Length of the rRNA sequence for RibosomialRNAGene.
}
\examples{
library(GenomicRanges)
library(Biostrings)
rRNA_example <- new("RibosomialRNAGene",
  geneID = "ENSG00000279457",
  symbol = "rRNA-5s",
  name = "5s ribosomial RNA",
  description = "rRNA component of the ribosome 5S subunit",
  structure = GRanges(
    seqnames = Rle("1"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("+")
  ),
  rRNAID = "RRNA0005",
  rRNASequence = RNAString("GCUCAGUUGGGAGAG")
)
rRNA_length <- lengthProduct(rRNA_example)
print(rRNA_length)
}
