% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_GeneFunctions.R
\name{lengthProduct,SmallNuclearRNAGene-method}
\alias{lengthProduct,SmallNuclearRNAGene-method}
\title{Length of the snRNA sequence for SmallNuclearRNAGene.}
\usage{
\S4method{lengthProduct}{SmallNuclearRNAGene}(gene)
}
\arguments{
\item{gene}{A SmallNuclearRNAGene object}
}
\value{
Integer representing the length of the snRNA sequence.
}
\description{
Length of the snRNA sequence for SmallNuclearRNAGene.
}
\examples{
library(GenomicRanges)
library(Biostrings)
snRNA_example <- new("SmallNuclearRNAGene",
  geneID = "ENSG00000234745",
  symbol = "snRNP-U1",
  name = "Small nuclear ribonucleoprotein U1",
  description = "snRNA component of the spliceosome",
  structure = GRanges(
    seqnames = Rle("Y"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("-")
  ),
  snRNAID = "SNR0001",
  snRNASequence = RNAString("GCCUUUAGCUCAAGCGGCUUAGAG")
)
snRNA_length <- lengthProduct(snRNA_example)
print(snRNA_length)
}
