% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_GeneFunctions.R
\name{lengthProduct,SRPRNAGene-method}
\alias{lengthProduct,SRPRNAGene-method}
\title{Length of the snRNA sequence for Signal recognition particle RNAGene.}
\usage{
\S4method{lengthProduct}{SRPRNAGene}(gene)
}
\arguments{
\item{gene}{A SRPRNAGene object}
}
\value{
Integer representing the length of the SRP RNA sequence.
}
\description{
Length of the snRNA sequence for Signal recognition particle RNAGene.
}
\examples{
library(GenomicRanges)
library(Biostrings)
SRP_example <- new("SRPRNAGene",
  geneID = "ENSG00000248527",
  symbol = "SRP-RNA",
  name = "Signal recognition particle RNA",
  description = "SRP RNA involved in targeting secretory proteins",
  structure = GRanges(
    seqnames = Rle("5"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("-")
  ),
  SRPRNAID = "SNR0001",
  SRPSequence = RNAString("GCCUUUAGCUCAAGCGGCUUAGAG")
)
SRP_length <- lengthProduct(SRP_example)
print(SRP_length)
}
