% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_GeneFunctions.R
\name{lengthProduct}
\alias{lengthProduct}
\title{Product length of a gene}
\usage{
lengthProduct(gene)
}
\arguments{
\item{gene}{The gene object from which the product length will be calculated.}
}
\value{
Returns the length of the gene product.
}
\description{
Product length of a gene
}
