% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{lncRNAID}
\alias{lncRNAID}
\alias{lncRNAID,LongNonCodingRNAGene-method}
\title{Accessor function for lncRNAID}
\usage{
lncRNAID(gene)

\S4method{lncRNAID}{LongNonCodingRNAGene}(gene)
}
\arguments{
\item{gene}{An instance of the LongNonCodingRNAGene class}
}
\value{
The lncRNAID of the gene.

The Gene object with updated lncRNAID
}
\description{
Accessor function for lncRNAID
}
\examples{
library(GenomicRanges)
library(Biostrings)
lnc_example <- new("LongNonCodingRNAGene",
  geneID = "ENSG00000228630",
  symbol = "HOTAIR",
  name = "HOX transcript antisense RNA",
  description = "LncRNA implicated in gene silencing.",
  structure = GRanges(
    seqnames = Rle("12"),
    ranges = IRanges(start = 53947302, end = 53948402),
    strand = Rle("-")
  ),
  lncRNAID = "NR_00316",
  RNASequence = RNAString("UGAGGUAGUAGGUU")
)
lncRNAID(lnc_example)
}
