% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{LongNonCodingRNAGene-class}
\alias{LongNonCodingRNAGene-class}
\title{Long non-coding Gene class}
\value{
Returns an object of class \code{LongNonCodingRNAGene}
}
\description{
Long non-coding RNA Gene Class
}
\details{
S4 class representing a long non-coding gene.
Inherits all slots from \code{\link{Gene}}.
}
\section{Slots}{

\describe{
\item{\code{lncRNAID}}{Unique lncRNA identifier}

\item{\code{RNASequence}}{Nucleotide sequence of lncRNA.}
}}

\examples{
library(GenomicRanges)
library(Biostrings)
lnc_example <- new("LongNonCodingRNAGene",
  geneID = "ENSG00000228630",
  symbol = "HOTAIR",
  name = "HOX transcript antisense RNA",
  description = "LncRNA implicated in gene silencing.",
  structure = GRanges(
    seqnames = Rle("12"),
    ranges = IRanges(start = 53947302, end = 53948402),
    strand = Rle("-")
  ),
  lncRNAID = "NR_00316",
  RNASequence = RNAString("UGAGGUAGUAGGUU")
)
print(lnc_example)
}
