% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_GeneConstructors.R
\name{LongNonCodingRNAGene}
\alias{LongNonCodingRNAGene}
\alias{createLongNonCodingRNAGene}
\title{Create a LongNonCodingRNAGene object}
\usage{
LongNonCodingRNAGene(
  geneID,
  symbol,
  name,
  description,
  structure,
  lncRNAID,
  RNASequence
)
}
\arguments{
\item{geneID}{Unique identifier for the gene.}

\item{symbol}{Standard symbol associated with the gene.}

\item{name}{Full name of the gene.}

\item{description}{Detailed description of the gene.}

\item{structure}{A GRanges object of the genomic structure of the gene.}

\item{lncRNAID}{Unique identifier for lncRNA.}

\item{RNASequence}{An RNAString containing the nucleotide sequence.}
}
\value{
A new LongNonCodingRNAGene object.
}
\description{
Constructor function for creating an instance of LongNonCodingRNAGene class.
}
\examples{
library(GenomicRanges)
library(Biostrings)
lnc_example <- new("LongNonCodingRNAGene",
  geneID = "ENSG00000228630",
  symbol = "HOTAIR",
  name = "HOX transcript antisense RNA",
  description = "LncRNA implicated in gene silencing.",
  structure = GRanges(
    seqnames = Rle("12"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("-")
  ),
  lncRNAID = "NR_00316",
  RNASequence = RNAString("UGAGGUAGUAGGUU")
)
print(lnc_example)

}
