% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_marker_names.R
\name{make_marker_names}
\alias{make_marker_names}
\title{Make genotypeR compliant marker names from the output of read_in_Master_SNPs_data function}
\usage{
make_marker_names(x)
}
\arguments{
\item{x}{Output of read_in_Master_SNPs_data}
}
\value{
A dataframe of GrandMasterSNPs markers with correct marker names
}
\description{
Make genotypeR compliant marker names from the output of read_in_Master_SNPs_data function
}
\examples{
\dontrun{
GrandMasterSNPs_markers  <- read_in_Master_SNPs_data("GrandMasterSNPs_output")
marker_names_GrandMasterSNPs_markers <- make_marker_names(GrandMasterSNPs_markers)
##If subset of markers needed
##use the sequenom output to subset the overall marker set from GrandMasterSNPs output
seq_test_data <- read_in_sequenom_data("Data/2_sequenom_plate_2_chip_8_dpseudo_plasticity_1_10_2017.csv")
markers_in_study <- test_data_marker_names[test_data_marker_names$marker_names\%in\%colnames(seq_test_data),]
}
}
\keyword{GrandMasterSNPs}
\keyword{associated}
\keyword{compliant}
\keyword{genotypeR}
\keyword{marker}
\keyword{markers}
\keyword{names}
\keyword{the}
\keyword{with}
