% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\docType{data}
\name{markers}
\alias{markers}
\title{Marker data produced with genotypeR}
\format{An object read in with \code{read_in_Master_SNPs_data}; see \code{\link[genotypeR]{read_in_Master_SNPs_data}}.}
\usage{
data(markers)
}
\description{
Data from a recombination plasticity experiment in 
Drosophila pseudoobscura. This data is provided to demonstrate the
the use of the genotypeR package
}
\examples{
data(markers)
head(markers)
colnames(markers)

}
\keyword{data}
\keyword{sets}
