% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{MicroRNAGene-class}
\alias{MicroRNAGene-class}
\title{MicroRNAGene}
\value{
Returns an object of class \code{MicroRNAGene}
}
\description{
Micro RNA Gene Class
}
\details{
S4 class representing microRNA gene.
Inherits all slots from \code{\link{Gene}}.
}
\section{Slots}{

\describe{
\item{\code{microRNAID}}{Unique microRNA identifier}

\item{\code{seedSequence}}{Nucleotide microRNA sequence}
}}

\examples{
library(GenomicRanges)
library(Biostrings)
miRNA_example <- new("MicroRNAGene",
  geneID = "ENSG00000207734",
  symbol = "MIR21",
  name = "MicroRNA 21",
  description = "miRNA involved in cancer regulation and expression.",
  structure = GRanges(
    seqnames = Rle("17"),
    ranges = IRanges(start = 29570, end = 29691),
    strand = Rle("+")
  ),
  microRNAID = "MI000077",
  seedSequence = RNAString("UAGCUUAUCAGAC")
)
print(miRNA_example)
}
