% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_GeneConstructors.R
\name{MicroRNAGene}
\alias{MicroRNAGene}
\alias{createMicroRNAGene}
\title{Create a MicroRNAGene object}
\usage{
MicroRNAGene(
  geneID,
  symbol,
  name,
  description,
  structure,
  microRNAID,
  seedSequence
)
}
\arguments{
\item{geneID}{Unique identifier for the gene.}

\item{symbol}{Standard symbol associated with the gene.}

\item{name}{Full name of the gene.}

\item{description}{Detailed description of the gene.}

\item{structure}{A GRanges object of the genomic structure of the gene.}

\item{microRNAID}{Unique identifier for microRNA.}

\item{seedSequence}{An RNAString containing the nucleotide sequence.}
}
\value{
A new MicroRNAGene object.
}
\description{
Constructor function for creating an instance of MicroRNAGene class.
}
\examples{
library(GenomicRanges)
library(Biostrings)
miRNA_example <- new("MicroRNAGene",
  geneID = "ENSG00000207734",
  symbol = "MIR21",
  name = "MicroRNA 21",
  description = "miRNA involved in cancer regulation and expression.",
  structure = GRanges(
    seqnames = Rle("17"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("+")
  ),
  microRNAID = "MI000077",
  seedSequence = RNAString("UAGCUUAUCAGAC")
)
print(miRNA_example)

}
