% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{microRNAID}
\alias{microRNAID}
\alias{microRNAID,MicroRNAGene-method}
\title{Accessor function for microRNAID}
\usage{
microRNAID(gene)

\S4method{microRNAID}{MicroRNAGene}(gene)
}
\arguments{
\item{gene}{An instance of the MicroRNAGene class}
}
\value{
The microRNAID of the gene.

The Gene object with updated microRNAID
}
\description{
Accessor function for microRNAID
}
\examples{
library(GenomicRanges)
library(Biostrings)
miRNA_example <- new("MicroRNAGene",
  geneID = "ENSG00000207734",
  symbol = "MIR21",
  name = "MicroRNA 21",
  description = "miRNA involved in cancer regulation and expression.",
  structure = GRanges(
    seqnames = Rle("17"),
    ranges = IRanges(start = 29570, end = 29691),
    strand = Rle("+")
  ),
  microRNAID = "MI000077",
  seedSequence = RNAString("UAGCUUAUCAGAC")
)
microRNAID(miRNA_example)
}
