% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{name}
\alias{name}
\alias{name,Gene-method}
\title{Accessor function for name}
\usage{
name(gene)

\S4method{name}{Gene}(gene)
}
\arguments{
\item{gene}{An instance of the Gene class.}
}
\value{
The name of the gene.

The Gene object with updated name
}
\description{
Accessor function for name
}
\examples{
library(GenomicRanges)
library(Biostrings)
pcg_example <- new("ProteinCodingGene",
  geneID = "ENSG00000139618",
  symbol = "BRCA2",
  name = "Breast cancer 2",
  description = "Gene implicated in breast cancer.",
  structure = GRanges(
    seqnames = Rle("17"),
    ranges = IRanges(start = 43044295, end = 43125482),
    strand = Rle("+")
  ),
  proteinID = "P51587",
  proteinSequence = AAString("MARKSLEMSIR")
)
name(pcg_example)
}
