% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{piRNAID}
\alias{piRNAID}
\alias{piRNAID,PiwiInteractingRNAGene-method}
\title{Accessor function for piRNAID}
\usage{
piRNAID(gene)

\S4method{piRNAID}{PiwiInteractingRNAGene}(gene)
}
\arguments{
\item{gene}{An instance of the PiwiInteractingRNAGene class}
}
\value{
The piRNAID of the gene.

The Gene object with updated piRNAID
}
\description{
Accessor function for piRNAID
}
\examples{
library(GenomicRanges)
library(Biostrings)
piRNA_example <- new("PiwiInteractingRNAGene",
  geneID = "ENSG00000241599",
  symbol = "PIWI001",
  name = "Piwi-interacting RNA 1",
  description = "Piwi-interacting RNA involved in gene regulation",
  structure = GRanges(
    seqnames = Rle("x"),
    ranges = IRanges(start = 1547627, end = 1577799),
    strand = Rle("+")
  ),
  piRNAID = "PIR0001",
  piSequence = RNAString("UGAGGUAGUAGGUU")
)
piRNAID(piRNA_example)
}
