% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{piSequence}
\alias{piSequence}
\alias{piSequence,PiwiInteractingRNAGene-method}
\title{Accessor function for piSequence}
\usage{
piSequence(gene)

\S4method{piSequence}{PiwiInteractingRNAGene}(gene)
}
\arguments{
\item{gene}{An instance of the PiwiInteractingRNAGene class}
}
\value{
The piSequence of the gene.

The Gene object with updated piSequence
}
\description{
Accessor function for piSequence
}
\examples{
library(GenomicRanges)
library(Biostrings)
piRNA_example <- new("PiwiInteractingRNAGene",
  geneID = "ENSG00000241599",
  symbol = "PIWI001",
  name = "Piwi-interacting RNA 1",
  description = "Piwi-interacting RNA involved in gene regulation",
  structure = GRanges(
    seqnames = Rle("x"),
    ranges = IRanges(start = 1547627, end = 1577799),
    strand = Rle("+")
  ),
  piRNAID = "PIR0001",
  piSequence = RNAString("UGAGGUAGUAGGUU")
)
piSequence(piRNA_example)
}
