% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_GeneConstructors.R
\name{PiwiInteractingRNAGene}
\alias{PiwiInteractingRNAGene}
\alias{createPiwiInteractingRNAGene}
\title{Create a PiwiInteractingRNAGene object}
\usage{
PiwiInteractingRNAGene(
  geneID,
  symbol,
  name,
  description,
  structure,
  piRNAID,
  piSequence
)
}
\arguments{
\item{geneID}{Unique identifier for the gene.}

\item{symbol}{Standard symbol associated with the gene.}

\item{name}{Full name of the gene.}

\item{description}{Detailed description of the gene.}

\item{structure}{A GRanges object of the genomic structure of the gene.}

\item{piRNAID}{Unique identifier for piRNA.}

\item{piSequence}{An RNAString containing the nucleotide sequence.}
}
\value{
A new PiwiInteractingRNAGene object.
}
\description{
Constructor function for creating an instance of PiwiInteractingRNAGene class
}
\examples{
library(GenomicRanges)
library(Biostrings)
piRNA_example <- new("PiwiInteractingRNAGene",
  geneID = "ENSG00000241599",
  symbol = "PIWI001",
  name = "Piwi-interacting RNA 1",
  description = "Piwi-interacting RNA involved in gene regulation",
  structure = GRanges(
    seqnames = Rle("x"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("+")
  ),
  piRNAID = "PIR0001",
  piSequence = RNAString("UGAGGUAGUAGGUU")
)
print(piRNA_example)

}
