% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{ProteinCodingGene-class}
\alias{ProteinCodingGene-class}
\title{ProteinCodingGene}
\value{
Returns an object of class \code{ProteinCodingGene}
}
\description{
Protein-Coding Gene Class
}
\details{
S4 class representing a protein-coding gene.
Inherits all slots from \code{\link{Gene}}.
}
\section{Slots}{

\describe{
\item{\code{proteinID}}{Unique protein identifier}

\item{\code{proteinSequence}}{Amino acid sequence (AAString)}
}}

\examples{
library(GenomicRanges)
library(Biostrings)
pcg_example <- new("ProteinCodingGene",
  geneID = "ENSG00000139618",
  symbol = "BRCA2",
  name = "Breast cancer 2",
  description = "Gene implicated in breast cancer.",
  structure = GRanges(
    seqnames = Rle("17"),
    ranges = IRanges(start = 43044295, end = 43125482),
    strand = Rle("+")
  ),
  proteinID = "P51587",
  proteinSequence = AAString("MARKSLEMSIR")
)
print(pcg_example)
}
