% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_GeneConstructors.R
\name{ProteinCodingGene}
\alias{ProteinCodingGene}
\alias{ProteinCodingGeneGene}
\alias{createProteinCodingGene}
\title{Create a ProteinCodingGene object}
\usage{
ProteinCodingGene(
  geneID,
  symbol,
  name,
  description,
  structure,
  proteinID,
  proteinSequence
)
}
\arguments{
\item{geneID}{Unique identifier for the gene.}

\item{symbol}{Standard symbol associated with the gene.}

\item{name}{Full name of the gene.}

\item{description}{Detailed description of the gene.}

\item{structure}{A GRanges object of the genomic structure of the gene.}

\item{proteinID}{Unique identifier for the protein.}

\item{proteinSequence}{Amino acid sequence of the protein.}
}
\value{
A new ProteinCodingGene object.
}
\description{
Constructor function for creating an instance of ProteinCodingGene class.
}
\examples{
library(GenomicRanges)
library(Biostrings)
pcg_example <- new("ProteinCodingGene",
  geneID = "ENSG00000139618",
  symbol = "BRCA2",
  name = "Breast cancer 2",
  description = "Gene implicated in breast cancer.",
  structure = GRanges(
    seqnames = Rle("17"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("+")
  ),
  proteinID = "P51587",
  proteinSequence = AAString("MVLSPADKTNVK")
)
pcg_example

}
