% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{proteinID}
\alias{proteinID}
\alias{proteinID,ProteinCodingGene-method}
\title{Accessor function for proteinID}
\usage{
proteinID(gene)

\S4method{proteinID}{ProteinCodingGene}(gene)
}
\arguments{
\item{gene}{An instance of the ProteinCodingGene class}
}
\value{
The proteinID of the gene.

The Gene object with updated proteinID
}
\description{
Accessor function for proteinID
}
\examples{
library(GenomicRanges)
library(Biostrings)
pcg_example <- new("ProteinCodingGene",
  geneID = "ENSG00000139618",
  symbol = "BRCA2",
  name = "Breast cancer 2",
  description = "Gene implicated in breast cancer.",
  structure = GRanges(
    seqnames = Rle("17"),
    ranges = IRanges(start = 43044295, end = 43125482),
    strand = Rle("+")
  ),
  proteinID = "P51587",
  proteinSequence = AAString("MARKSLEMSIR")
)
proteinID(pcg_example)
}
