% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{RibosomialRNAGene-class}
\alias{RibosomialRNAGene-class}
\title{RibosomialRNAGene}
\value{
Returns an object of class \code{RibosomialRNAGene}
}
\description{
Ribosomial RNA Gene Class
}
\details{
S4 class representing microRNA gene.
Inherits all slots from \code{\link{Gene}}.
}
\section{Slots}{

\describe{
\item{\code{rRNAID}}{Unique microRNA identifier}

\item{\code{rRNASequence}}{Nucleotide microRNA sequence}
}}

\examples{
library(GenomicRanges)
library(Biostrings)
rRNA_example <- new("RibosomialRNAGene",
  geneID = "ENSG00000279457",
  symbol = "rRNA-5s",
  name = "5s ribosomial RNA",
  description = "rRNA component of the ribosome 5S subunit",
  structure = GRanges(
    seqnames = Rle("1"),
    ranges = IRanges(start = 3413204, end = 3413561),
    strand = Rle("+")
  ),
  rRNAID = "RRNA0005",
  rRNASequence = RNAString("GCUCAGUUGGGAGAG")
)
rRNASequence(rRNA_example)
print(rRNA_example)
}
