% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_GeneConstructors.R
\name{RibosomialRNAGene}
\alias{RibosomialRNAGene}
\alias{createRibosomialRNAGene}
\title{Create a RibosomialRNAGene object}
\usage{
RibosomialRNAGene(
  geneID,
  symbol,
  name,
  description,
  structure,
  rRNAID,
  rRNASequence
)
}
\arguments{
\item{geneID}{Unique identifier for the gene.}

\item{symbol}{Standard symbol associated with the gene.}

\item{name}{Full name of the gene.}

\item{description}{Detailed description of the gene.}

\item{structure}{A GRanges object of the genomic structure of the gene.}

\item{rRNAID}{Unique identifier for rRNA.}

\item{rRNASequence}{An RNAString containing the nucleotide sequence.}
}
\value{
A new RibosomialRNAGene object.
}
\description{
Constructor function for creating an instance of RibosomialRNAGene class.
}
\examples{
library(GenomicRanges)
library(Biostrings)
rRNA_example <- new("RibosomialRNAGene",
  geneID = "ENSG00000279457",
  symbol = "rRNA-5s",
  name = "5s ribosomial RNA",
  description = "rRNA component of the ribosome 5S subunit",
  structure = GRanges(
    seqnames = Rle("1"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("+")
  ),
  rRNAID = "RRNA0005",
  rRNASequence = RNAString("GCUCAGUUGGGAGAG")
)
print(rRNA_example)

}
