% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{RNASequence}
\alias{RNASequence}
\alias{RNASequence,LongNonCodingRNAGene-method}
\title{Accessor function for RNASequence}
\usage{
RNASequence(gene)

\S4method{RNASequence}{LongNonCodingRNAGene}(gene)
}
\arguments{
\item{gene}{An instance of the LongNonCodingRNAGene class}
}
\value{
The RNASequence of the gene.

The Gene object with updated RNASequence
}
\description{
Accessor function for RNASequence
}
\examples{
library(GenomicRanges)
library(Biostrings)
lnc_example <- new("LongNonCodingRNAGene",
  geneID = "ENSG00000228630",
  symbol = "HOTAIR",
  name = "HOX transcript antisense RNA",
  description = "LncRNA implicated in gene silencing.",
  structure = GRanges(
    seqnames = Rle("12"),
    ranges = IRanges(start = 53947302, end = 53948402),
    strand = Rle("-")
  ),
  lncRNAID = "NR_00316",
  RNASequence = RNAString("UGAGGUAGUAGGUU")
)
RNASequence(lnc_example)
}
