% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{rRNASequence}
\alias{rRNASequence}
\alias{rRNASequence,RibosomialRNAGene-method}
\title{Accessor function for rRNASequence}
\usage{
rRNASequence(gene)

\S4method{rRNASequence}{RibosomialRNAGene}(gene)
}
\arguments{
\item{gene}{An instance of the RibosomialRNAGene class}
}
\value{
The rRNASequence of the gene.

The Gene object with updated rRNASequence
}
\description{
Accessor function for rRNASequence
}
\examples{
library(GenomicRanges)
library(Biostrings)
rRNA_example <- new("RibosomialRNAGene",
  geneID = "ENSG00000279457",
  symbol = "rRNA-5s",
  name = "5s ribosomial RNA",
  description = "rRNA component of the ribosome 5S subunit",
  structure = GRanges(
    seqnames = Rle("1"),
    ranges = IRanges(start = 3413204, end = 3413561),
    strand = Rle("+")
  ),
  rRNAID = "RRNA0005",
  rRNASequence = RNAString("GCUCAGUUGGGAGAG")
)
rRNASequence(rRNA_example)
}
