% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{seedSequence}
\alias{seedSequence}
\alias{seedSequence,MicroRNAGene-method}
\title{Accessor function for seedSequence}
\usage{
seedSequence(gene)

\S4method{seedSequence}{MicroRNAGene}(gene)
}
\arguments{
\item{gene}{An instance of the MicroRNAGene class}
}
\value{
The seedSequence of the gene.

The Gene object with updated seedSequence
}
\description{
Accessor function for seedSequence
}
\examples{
library(GenomicRanges)
library(Biostrings)
miRNA_example <- new("MicroRNAGene",
  geneID = "ENSG00000207734",
  symbol = "MIR21",
  name = "MicroRNA 21",
  description = "miRNA involved in cancer regulation and expression.",
  structure = GRanges(
    seqnames = Rle("17"),
    ranges = IRanges(start = 29570, end = 29691),
    strand = Rle("+")
  ),
  microRNAID = "MI000077",
  seedSequence = RNAString("UAGCUUAUCAGAC")
)
seedSequence(miRNA_example)
}
