% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{SmallNuclearRNAGene-class}
\alias{SmallNuclearRNAGene-class}
\title{SmallNuclearRNAGene}
\value{
Returns an object of class \code{SmallNuclearRNAGene}
}
\description{
Small Nuclear RNA Gene Class
}
\details{
This S4 class represents smallnuclearRNA gene
Inherits all slots from \code{\link{Gene}}.
}
\section{Slots}{

\describe{
\item{\code{snRNAID}}{Unique snRNA identifier}

\item{\code{snRNASequence}}{Nucleotide snRNA sequence}
}}

\examples{
library(GenomicRanges)
library(Biostrings)
snRNA_example <- new("SmallNuclearRNAGene",
  geneID = "ENSG00000234745",
  symbol = "snRNP-U1",
  name = "Small nuclear ribonucleoprotein U1",
  description = "snRNA component of the spliceosome",
  structure = GRanges(
    seqnames = Rle("Y"),
    ranges = IRanges(start = 26357, end = 29428),
    strand = Rle("-")
  ),
  snRNAID = "SNR0001",
  snRNASequence = RNAString("GCCUUUAGCUCAAGCGGCUUAGAG")
)
print(snRNA_example)
}
