% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_GeneConstructors.R
\name{SmallNuclearRNAGene}
\alias{SmallNuclearRNAGene}
\alias{createSmallNuclearRNAGene}
\title{Create a SmallNuclearRNAGene object}
\usage{
SmallNuclearRNAGene(
  geneID,
  symbol,
  name,
  description,
  structure,
  snRNAID,
  snRNASequence
)
}
\arguments{
\item{geneID}{Unique identifier for the gene.}

\item{symbol}{Standard symbol associated with the gene.}

\item{name}{Full name of the gene.}

\item{description}{Detailed description of the gene.}

\item{structure}{A GRanges object of the genomic structure of the gene.}

\item{snRNAID}{Unique identifier for snRNA.}

\item{snRNASequence}{An RNAString containing the nucleotide sequence.}
}
\value{
A new SmallNuclearRNAGene object.
}
\description{
Constructor function for creating an instance of SmallNuclearRNAGene class.
}
\examples{
library(GenomicRanges)
library(Biostrings)
snRNA_example <- new("SmallNuclearRNAGene",
  geneID = "ENSG00000234745",
  symbol = "snRNP-U1",
  name = "Small nuclear ribonucleoprotein U1",
  description = "snRNA component of the spliceosome",
  structure = GRanges(
    seqnames = Rle("Y"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("-")
  ),
  snRNAID = "SNR0001",
  snRNASequence = RNAString("GCCUUUAGCUCAAGCGGCUUAGAG")
)
print(snRNA_example)

}
