% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_sequenom_df.R
\name{sort_sequenom_df}
\alias{sort_sequenom_df}
\title{Sequenom Data frame Sort}
\usage{
sort_sequenom_df(Sequenom_Data2Sort, sort_char = "chr|contig")
}
\arguments{
\item{Sequenom_Data2Sort}{data frame to sort produced with the
genotypeR package}

\item{sort_char}{is the character string output by the PERL pipeline in the marker design phase
(i.e., chr 1000 1050 AAA[A/T]GTC; the chr is the sort_char. Defaults to chr or contig.}
}
\value{
A sorted data frame suited for the genotypeR package
}
\description{
This function sorts Sequenom Data at the read-in stage.
}
\examples{

##sequenom_data <- read_in_sequenom_data("your.csv")
##or
data(genotypes_data)
sort_sequenom_df(genotypes_data)

}
\keyword{sequenom}
\keyword{sort}
