% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{SRPRNAGene-class}
\alias{SRPRNAGene-class}
\title{SRPRNAGene}
\value{
Returns an object of class \code{SRPRNAGene}
}
\description{
Signal Recognition Particle RNA Gene Class
}
\details{
This S4 class represents smallnuclearRNA gene
Inherits all slots from \code{\link{Gene}}.
}
\section{Slots}{

\describe{
\item{\code{SRPRNAID}}{Unique snRNA identifier}

\item{\code{SRPSequence}}{Nucleotide snRNA sequence}
}}

\examples{
library(GenomicRanges)
library(Biostrings)
SRP_example <- new("SRPRNAGene",
  geneID = "ENSG00000248527",
  symbol = "SRP-RNA",
  name = "Signal recognition particle RNA",
  description = "SRP RNA involved in targeting secretory proteins",
  structure = GRanges(
    seqnames = Rle("5"),
    ranges = IRanges(start = 112345, end = 112423),
    strand = Rle("-")
  ),
  SRPRNAID = "SNR0001",
  SRPSequence = RNAString("GCCUUUAGCUCAAGCGGCUUAGAG")
)
print(SRP_example)
}
