% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_GeneConstructors.R
\name{SRPRNAGene}
\alias{SRPRNAGene}
\alias{createSRPRNAGene}
\title{Create a SRPRNAGene object}
\usage{
SRPRNAGene(geneID, symbol, name, description, structure, SRPRNAID, SRPSequence)
}
\arguments{
\item{geneID}{Unique identifier for the gene.}

\item{symbol}{Standard symbol associated with the gene.}

\item{name}{Full name of the gene.}

\item{description}{Detailed description of the gene.}

\item{structure}{A GRanges object of the genomic structure of the gene.}

\item{SRPRNAID}{Unique identifier for SRPRNA.}

\item{SRPSequence}{An RNAString containing the nucleotide sequence.}
}
\value{
A new SRPRNAGene object.
}
\description{
Constructor function for creating an instance of SRPRNAGene class.
}
\examples{
library(GenomicRanges)
library(Biostrings)
SRP_example <- new("SRPRNAGene",
  geneID = "ENSG00000248527",
  symbol = "SRP-RNA",
  name = "Signal recognition particle RNA",
  description = "SRP RNA involved in targeting secretory proteins",
  structure = GRanges(
    seqnames = Rle("5"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("-")
  ),
  SRPRNAID = "SNR0001",
  SRPSequence = RNAString("GCCUUUAGCUCAAGCGGCUUAGAG")
)
print(SRP_example)

}
