% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{SRPRNAID}
\alias{SRPRNAID}
\alias{SRPRNAID,SRPRNAGene-method}
\title{Accessor function for SRPRNAID}
\usage{
SRPRNAID(gene)

\S4method{SRPRNAID}{SRPRNAGene}(gene)
}
\arguments{
\item{gene}{An instance of the SRPRNAGene class}
}
\value{
The SRPRNAID of the gene.

The Gene object with updated SRPRNAID
}
\description{
Accessor function for SRPRNAID
}
\examples{
library(GenomicRanges)
library(Biostrings)
SRP_example <- new("SRPRNAGene",
  geneID = "ENSG00000248527",
  symbol = "SRP-RNA",
  name = "Signal recognition particle RNA",
  description = "SRP RNA involved in targeting secretory proteins",
  structure = GRanges(
    seqnames = Rle("5"),
    ranges = IRanges(start = 112345, end = 112423),
    strand = Rle("-")
  ),
  SRPRNAID = "SNR0001",
  SRPSequence = RNAString("GCCUUUAGCUCAAGCGGCUUAGAG")
)
SRPRNAID(SRP_example)
}
