% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{SRPSequence}
\alias{SRPSequence}
\alias{SRPSequence,SRPRNAGene-method}
\title{Accessor function for SRPSequence}
\usage{
SRPSequence(gene)

\S4method{SRPSequence}{SRPRNAGene}(gene)
}
\arguments{
\item{gene}{An instance of the SRPRNAGene class}
}
\value{
The SRPSequence of the gene.

The Gene object with updated SRPSequence
}
\description{
Accessor function for SRPSequence
}
\examples{
library(GenomicRanges)
library(Biostrings)
SRP_example <- new("SRPRNAGene",
  geneID = "ENSG00000248527",
  symbol = "SRP-RNA",
  name = "Signal recognition particle RNA",
  description = "SRP RNA involved in targeting secretory proteins",
  structure = GRanges(
    seqnames = Rle("5"),
    ranges = IRanges(start = 112345, end = 112423),
    strand = Rle("-")
  ),
  SRPRNAID = "SNR0001",
  SRPSequence = RNAString("GCCUUUAGCUCAAGCGGCUUAGAG")
)
SRPSequence(SRP_example)
}
