% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetChromosome.R
\name{subsetChromosome}
\alias{subsetChromosome}
\title{Subset genotypeR object by chromosome}
\usage{
subsetChromosome(aa, chromosome)
}
\arguments{
\item{aa}{genotypeR object before binary coding}

\item{chromosome}{which chromosome to pull out (e.g., "chr2")}
}
\value{
A subsetted genotypeR object
}
\description{
Subset genotypeR object by chromosome
}
\examples{
\dontrun{
chromosome_subset <- subsetChromosome(aa, chromosome)
}##subset function
}
\keyword{by}
\keyword{chromosome}
\keyword{genotypeR}
\keyword{object}
\keyword{subset}
