% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{TransferRNAGene-class}
\alias{TransferRNAGene-class}
\title{TransferRNAGene}
\value{
Returns an object of class \code{TransferRNAGene}
}
\description{
Transfer RNA Gene Class
}
\details{
This S4 class represents transferRNA gene
Inherits all slots from \code{\link{Gene}}.
}
\section{Slots}{

\describe{
\item{\code{tRNAID}}{Unique tRNA identifier}

\item{\code{tRNASequence}}{Nucleotide tRNA sequence}
}}

\examples{
library(GenomicRanges)
library(Biostrings)
tRNA_example <- new("TransferRNAGene",
  geneID = "ENSG00000201145",
  symbol = "tRNA-Lys",
  name = "Transfer RNA lysine",
  description = "tRNA involved in protein synthesis",
  structure = GRanges(
    seqnames = Rle("Y"),
    ranges = IRanges(start = 26357, end = 29428),
    strand = Rle("-")
  ),
  tRNAID = "TRNA023",
  tRNASequence = RNAString("GCCUUUAGCUCAGUUGGGAGAG")
)
tRNASequence(tRNA_example)
print(tRNA_example)
}
