% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_GeneConstructors.R
\name{TransferRNAGene}
\alias{TransferRNAGene}
\alias{createTransferRNAGene}
\title{Create a TransferRNAGene object}
\usage{
TransferRNAGene(
  geneID,
  symbol,
  name,
  description,
  structure,
  tRNAID,
  tRNASequence
)
}
\arguments{
\item{geneID}{Unique identifier for the gene.}

\item{symbol}{Standard symbol associated with the gene.}

\item{name}{Full name of the gene.}

\item{description}{Detailed description of the gene.}

\item{structure}{A GRanges object of the genomic structure of the gene.}

\item{tRNAID}{Unique identifier for tRNA.}

\item{tRNASequence}{An RNAString containing the nucleotide sequence.}
}
\value{
A new TransferRNAGene object.
}
\description{
Constructor function for creating an instance of TransferRNAGene class.
}
\examples{
library(GenomicRanges)
library(Biostrings)
tRNA_example <- new("TransferRNAGene",
  geneID = "ENSG00000201145",
  symbol = "tRNA-Lys",
  name = "Transfer RNA lysine",
  description = "tRNA involved in protein synthesis",
  structure = GRanges(
    seqnames = Rle("Y"),
    ranges = IRanges(start = 500, end = 600),
    strand = Rle("-")
  ),
  tRNAID = "TRNA023",
  tRNASequence = RNAString("GCCUUUAGCUCAGUUGGGAGAG")
)
print(tRNA_example)

}
