% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{tRNAID}
\alias{tRNAID}
\alias{tRNAID,TransferRNAGene-method}
\title{Accessor function for tRNAID}
\usage{
tRNAID(gene)

\S4method{tRNAID}{TransferRNAGene}(gene)
}
\arguments{
\item{gene}{An instance of the TransferRNAGene class}
}
\value{
The tRNAID of the gene.

The Gene object with updated tRNAID
}
\description{
Accessor function for tRNAID
}
\examples{
library(GenomicRanges)
library(Biostrings)
tRNA_example <- new("TransferRNAGene",
  geneID = "ENSG00000201145",
  symbol = "tRNA-Lys",
  name = "Transfer RNA lysine",
  description = "tRNA involved in protein synthesis",
  structure = GRanges(
    seqnames = Rle("Y"),
    ranges = IRanges(start = 26357, end = 29428),
    strand = Rle("-")
  ),
  tRNAID = "TRNA023",
  tRNASequence = RNAString("GCCUUUAGCUCAGUUGGGAGAG")
)
tRNAID(tRNA_example)
}
