% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_GeneAccessors.R
\name{tRNASequence}
\alias{tRNASequence}
\alias{tRNASequence,TransferRNAGene-method}
\title{Accessor function for tRNASequence}
\usage{
tRNASequence(gene)

\S4method{tRNASequence}{TransferRNAGene}(gene)
}
\arguments{
\item{gene}{An instance of the TransferRNAGene class}
}
\value{
The tRNASequence of the gene.

The Gene object with updated tRNASequence
}
\description{
Accessor function for tRNASequence
}
\examples{
library(GenomicRanges)
library(Biostrings)
tRNA_example <- new("TransferRNAGene",
  geneID = "ENSG00000201145",
  symbol = "tRNA-Lys",
  name = "Transfer RNA lysine",
  description = "tRNA involved in protein synthesis",
  structure = GRanges(
    seqnames = Rle("Y"),
    ranges = IRanges(start = 26357, end = 29428),
    strand = Rle("-")
  ),
  tRNAID = "TRNA023",
  tRNASequence = RNAString("GCCUUUAGCUCAGUUGGGAGAG")
)
tRNASequence(tRNA_example)
}
