% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO.R
\name{CO}
\alias{CO}
\title{Count crossovers per individual}
\usage{
CO(indata, naive = FALSE)
}
\value{
data frame of COs counted per individual
}
\description{
Count crossovers per individual
}
\examples{
\dontrun{
CO_df <- CO(indata, naive=FALSE)
}##subset function

}
\keyword{count}
\keyword{crossovers}
\keyword{genotypeR}
