% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2qtl_table.R
\name{convert2qtl_table}
\alias{convert2qtl_table}
\title{write out table for import into rqtl}
\usage{
convert2qtl_table(genotypeR_object,
  temp_cross_for_qtl = "temp_cross_for_qtl.csv")
}
\value{
table to disk for input into rqtl
}
\description{
write out table for import into rqtl
}
\examples{
\dontrun{
convert2qtl_table(counted_COs, temp_cross_for_qtl="~/Desktop/temp.csv")
}##subset function

}
\keyword{genotypeR,}
\keyword{rqtl}
