% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_CO.R
\name{count_CO}
\alias{count_CO}
\title{Internal function to remove search and remove columns based on names}
\usage{
count_CO(data, naive = FALSE)
}
\value{
genotypeR object with counted crossovers
}
\description{
Internal function to remove search and remove columns based on names
}
\examples{
\dontrun{
genotypeR_object_with_counted_COs_slot_added <- count_CO(data, naive=FALSE)
}##subset function

}
\keyword{Count}
\keyword{Crossovers}
\keyword{a}
\keyword{genotypeR}
\keyword{object}
\keyword{of}
