% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{class}
\name{genotypeR-class}
\alias{genotypeR-class}
\alias{crossOveRs}
\alias{genotypeR-class}
\title{Class genotypeR.}
\usage{
genotypes(object, ...)
}
\description{
Class \code{genotypeR} defines a gas sensor device.

Method genotypes.
}
