% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{class}
\name{genotypeR-class}
\alias{genotypeR-class}
\alias{genotypeR}
\title{Class genotypeR.}
\description{
Class \code{genotypeR} defines a gas sensor device.
}
\section{Slots}{

\describe{
\item{\code{genotypes}}{is a dataframe of genotypes}

\item{\code{impossible_genotype}}{is a vector}

\item{\code{binary_genotypes}}{is a dataframe of numeric coded genotypes}

\item{\code{counted_crossovers}}{is a dataframe of counted crossovers}
}}

