% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{methods}
\name{genotypes}
\alias{genotypes}
\alias{impossible_genotype}
\alias{counted_crossovers}
\alias{binary_genotypes<-}
\alias{genotypes<-}
\alias{counted_crossovers<-}
\alias{show}
\title{Method genotypes.}
\usage{
genotypes(object, ...)

impossible_genotype(object, ...)

counted_crossovers(object, ...)

binary_genotypes(object) <- value

genotypes(object) <- value

counted_crossovers(object) <- value

\S4method{show}{genotypeR}(object)
}
\description{
Method genotypes.

Method impossible_genotype.

Method counted_crossovers.

Method binary_genotypes<-.

Method genotypes<-.

Method counted_crossovers<-.

Method show.
}
