% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_df_subset.R
\name{grep_df_subset}
\alias{grep_df_subset}
\title{Internal function to remove search and remove columns based on names}
\usage{
grep_df_subset(x, toMatch)
}
\arguments{
\item{x}{dataframe where columns are to be removed}

\item{toMatch}{vector of characters to remove from x}
}
\value{
A subsetted genotypeR object
}
\description{
\code{grep_df_subset} is an interal function that subsets
a dataframe based on supplied pattern. This function is
provided incase it is found useful.
}
\examples{
\dontrun{
data_frame_columns_removed <- grep_df_subset(x, toMatch)
}##subset function
}
\keyword{columns}
\keyword{matching}
\keyword{remove}
