% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_df_subset.R
\name{grep_df_subset}
\alias{grep_df_subset}
\title{Internal function to remove search and remove columns based on names}
\usage{
grep_df_subset(x, toMatch)
}
\value{
A subsetted genotypeR object
}
\description{
Internal function to remove search and remove columns based on names
}
\examples{
\dontrun{
data_frame_columns_removed <- grep_df_subset(x, toMatch)
}##subset function
}
\keyword{columns}
\keyword{matching}
\keyword{remove}
