% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Master_SNPs_Data.R
\name{read_in_Master_SNPs_data}
\alias{read_in_Master_SNPs_data}
\title{Read in GrandMasterSNPs output}
\usage{
read_in_Master_SNPs_data(x, ...)
}
\arguments{
\item{x}{This is a tab delimited text file from GrandMasterSNPs Perl program}

\item{...}{Other arguments passed to the function}
}
\value{
A data frame of GrandMasterSNPs markers
}
\description{
This reads in single nucleotide polymorphism markers generated
by the GrandMasterSNPs Perl program.
}
\examples{
\dontrun{
GrandMasterSNPs_markers  <- read_in_Master_SNPs_data("GrandMasterSNPs_output")
}
}
\keyword{GrandMasterSNPs}
\keyword{read}
