% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_sequenom_data.R
\name{read_in_sequenom_data}
\alias{read_in_sequenom_data}
\title{Read in Sequenom Data}
\usage{
read_in_sequenom_data(x, ...)
}
\arguments{
\item{x}{This is a csv formated Genotypes tab of exported sequenom data that you would like to read in.}

\item{...}{Other arguments passed to the function}
}
\value{
A dataframe suited for the genotypeR package
}
\description{
This function is a wrapper function around read.csv in order to genotype data provide data compatible with the genotypeR package.
}
\examples{
\dontrun{
sequenom_data <- read_in_sequenom_data("your.csv")
}
}
\keyword{read}
\keyword{sequenom}
