% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ref_Alt_Table.R
\name{Ref_Alt_Table}
\alias{Ref_Alt_Table}
\title{Make reference/alternate allelle table from make_marker_names output}
\usage{
Ref_Alt_Table(markers_in_study)
}
\arguments{
\item{markers_in_study}{make_marker_names output}
}
\value{
A dataframe of Ref/Alt genotypes
}
\description{
\code{Ref_Alt_Table} makes the ref/alt table used
in for proper genotype codeing and
QA/QC \code{initialize_genotypeR_data}.
}
\examples{
\dontrun{
data(markers)
markers_in_study <- make_marker_names(markers)
genotype_table <- Ref_Alt_Table(markers_in_study = markers_in_study)
}
}
\keyword{make_marker_names}
