% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequenom_Markers_v2.R
\name{SequenomMarkers}
\alias{SequenomMarkers}
\title{R wrapper script to run Sequenom Marker design pipeline}
\usage{
SequenomMarkers(vcf1, vcf2, outdir)
}
\arguments{
\item{vcf1}{this is an uncompressed vcf file (Ref allele)}

\item{vcf2}{this is an uncompressed vcf file (Alt allele)}

\item{outdir}{this is where the tab-delimited extended bed file will
be written}
}
\value{
SequenomMarker design into "outdir"
}
\description{
This function designs Sequenom markers.
}
\examples{
\dontrun{
SequenomMarkers("/home/ssefick/R/x86_64-redhat-linux-gnu-library/3.3/genotypeR/SequenomMarkers/FS14_test.vcf", "/home/ssefick/R/x86_64-redhat-linux-gnu-library/3.3/genotypeR/SequenomMarkers/FS16_test.vcf", "/home/ssefick/Desktop/Stevison_PostDoc/Projects/genotypeR/inst/SequenomMarkers/test_dir")
}
}
\keyword{SequenomMarker}
