% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_sequenom_df.R
\name{sort_sequenom_df}
\alias{sort_sequenom_df}
\title{Sequenom Dataframe Sort}
\usage{
sort_sequenom_df(Sequenom_Data2Sort)
}
\arguments{
\item{Sequenom_Data2Sort}{dataframe to sort produced with the
genotypeR package}
}
\value{
A sorted dataframe suited for the genotypeR package
}
\description{
This function sorts Sequenom Data at the read-in stage.
}
\examples{
\dontrun{
##sequenom_data <- read_in_sequenom_data("your.csv")
##or
data(genotypes_data)
sort_sequenom_df(genotypes_data)
}
}
\keyword{sequenom}
\keyword{sort}
