% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_sequenom_df.R
\name{sort_sequenom_df}
\alias{sort_sequenom_df}
\title{Sequenom Dataframe Sort}
\usage{
sort_sequenom_df(Sequenom_Data2Sort)
}
\value{
A sorted dataframe suited for the genotypeR package
}
\description{
This function sorts Sequenom Data at the read-in stage.
}
\examples{
\dontrun{
sequenom_data <- read_in_sequenom_data("your.csv")
sort_sequenom_df(sequenom_data)
}
}
\keyword{sequenom}
\keyword{sort}
