% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_one_two_coding.R
\name{zero_one_two_coding}
\alias{zero_one_two_coding}
\title{Code genotypes as 0, 1, 2}
\usage{
zero_one_two_coding(genotype_warnings_passthrough, genotype_table)
}
\arguments{
\item{genotype_warnings_passthrough}{is a genotypeR object that has been processed by BC_Genotype_Warnings
with output="pass_through"}

\item{genotype_table}{is a dataframe produced with Ref_Alt_Table}
}
\value{
A dataframe of 0, 1, and 2 coded genotypes as a slot in the input
}
\description{
\code{zero_one_two_coding} code homozygous reference as 0, heterozygous as 1, and homozygous alternate as 2 using a genotypeR object created with \code{initialize_genotypeR_data} with the pass_through argument.
}
\examples{
\dontrun{
genotype_warnings_passthrough <- BC_Genotype_Warnings(seq_data =
seq_test_data, genotype_table = genotype_table, output="pass_though")
bb <- zero_one_two_coding(genotype_warnings_passthrough = bb,
genotype_table = genotype_table)
}
}
\keyword{0}
\keyword{1}
\keyword{2}
\keyword{Alt}
\keyword{Heterozygous,}
\keyword{Homozygous}
\keyword{Ref,}
\keyword{and}
\keyword{as}
\keyword{code}
\keyword{genotypes}
